FlashData, PanelDue: moved flashdata stuff into separate module
